Refactored app.rb to handle persons
